% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NSUM.R
\name{print.NSUM}
\alias{print.NSUM}
\title{Other functions}
\usage{
\method{print}{NSUM}(x, ...)
}
\arguments{
\item{x}{object returned by `nsum.mcmc()`}

\item{...}{other arguments to/from other methods}
}
\description{
Other functions
}
