% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NSUM.R
\name{summary.NSUM}
\alias{summary.NSUM}
\alias{print.summary.NSUM}
\title{Summarise NSUM model fits}
\usage{
\method{summary}{NSUM}(object, ...)

\method{print}{summary.NSUM}(x, ...)
}
\arguments{
\item{object}{object of class "`NSUM`" as returned by [nsum.mcmc()]}

\item{...}{other arguments to/from other methods, currently ignored}

\item{x}{object of class "`NSUM.summary` as returned by `summary.NSUM()`}
}
\value{
The function `summary.NSUM` returns a list with as many elements as
  `object` has, but elements corresponding to estimated parameters are
  summarised with [summary.default()].
}
\description{
[summary()] method for models fitted with [nsum.mcmc()].
}
